requirements.txt
streamlit
langchain
langchain-community
PyPDF2
tiktoken
uv
pytest
chromadb
sentence-transformers
python-dotenv
langchain-huggingface
langchain-openai
fastapi
redis
uvicorn
requests
pydantic
langchain-text-splitters
python-multipart
SQLAlchemy
psycopg2-binary
faiss-cpu
celery